+++ libraries/matchers.rb
@@ -48,4 +48,20 @@ if defined?(ChefSpec)
   def delete_influxdb_config(config)
     ChefSpec::Matchers::ResourceMatcher.new(:influxdb_config, :delete, config)
   end
+
+  def install_influxdb_install(package)
+    ChefSpec::Matchers::ResourceMatcher.new(:influxdb_install, :install, package)
+  end
+
+  def remove_influxdb_install(package)
+    ChefSpec::Matchers::ResourceMatcher.new(:influxdb_install, :remove, package)
+  end
+
+  def create_influxdb_continuous_query(continuous_query)
+    ChefSpec::Matchers::ResourceMatcher.new(:influxdb_continuous_query, :create, continuous_query)
+  end
+
+  def delete_influxdb_continuous_query(continuous_query)
+    ChefSpec::Matchers::ResourceMatcher.new(:influxdb_continuous_query, :delete, continuous_query)
+  end
 end